fib0

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,14 +7,14 @@
 int main(void) 
 {
 int count;
-long long int fib0 = 1, fib1 =2, prevfeb;
-printf("%lld, %lld", fib0, fib1);
+long int fib0 = 1, fib1 =2, prevfeb;
+printf("%ld, %ld", fib0, fib1);
 
 for (count = 0; count < 48; count ++)
 {
 prevfeb = fib1;
 fib1 = fib0 + fib1;
-printf(", %lld", fib1);
+printf(", %ld", fib1);
 fib0 = prevfeb;
 }
 printf("\n");
